+++ CumulusMX/EcowittHttpApiStation.cs
@@ -450,7 +450,7 @@ namespace CumulusMX
 				}
 			}
 
-			return null;
+			return string.Empty;
 		}
 
 		public override string GetEcowittVideoUrl()
@@ -468,7 +468,7 @@ namespace CumulusMX
 				}
 			}
 
-			return null;
+			return string.Empty;
 		}
 
 		private async Task GetSensorIds()