+++ CumulusMX/GW1000Station.cs
@@ -339,7 +339,7 @@ namespace CumulusMX
 				}
 			}
 
-			return null;
+			return string.Empty;
 		}
 
 		public override string GetEcowittVideoUrl()
@@ -357,7 +357,7 @@ namespace CumulusMX
 				}
 			}
 
-			return null;
+			return string.Empty;
 		}